+++ apps/wallet/frontend/src/routes/allocations.tsx
@@ -4,10 +4,12 @@ import React from 'react';
 import CreateAllocation from '../components/CreateAllocation';
 import { Box } from '@mui/material';
 import ListAllocationRequests from '../components/ListAllocationRequests';
+import ListAllocations from '../components/ListAllocations';
 
 const Allocations: React.FC = () => (
   <Box marginTop={4}>
     <ListAllocationRequests />
+    <ListAllocations />
     <CreateAllocation />
   </Box>
 );
